refactor error handling code out of application controller and into concern
